Improve test system
Add 'cleanup' make target

+++ travis/run-test.sh
@@ -6,25 +6,36 @@
 
 run_tests()
 {
+dir="$1"
+
 _ret=0
 
 echo
-echo "=============== Running tests in $PWD"
+echo "=============== Running tests in <base>/$dir"
 echo
 
+cd $BASE/$dir
 phpize || return 1
 ./configure --quiet || return 1
-make -k clean all CFLAGS='-g -Wall' || return 1
-
-# Remove this when the valgrind options in run-tests.php are fixed
-sed 's/--vex-iropt-register-updates=allregs-at-mem-access/--vex-iropt-precise-memory-exns=yes/g' <run-tests.php >r
-rm -f run-tests.php
-cp r run-tests.php
+make -k clean install CFLAGS='-g -Wall' || return 1
 
 # -m : test using valgrind
 # -q : No interaction
 
-TEST_PHP_ARGS='-m -q' make test
+TEST_PHP_ARGS='-q'
+if [ -n "$USE_VALGRIND" ] ; then
+	TEST_PHP_ARGS="-m $TEST_PHP_ARGS"
+
+	# Remove this when the valgrind options in run-tests.php are fixed
+	if [ "`valgrind --version`" = 'valgrind-3.7.0' ] ; then
+		sed 's/--vex-iropt-register-updates=allregs-at-mem-access/--vex-iropt-precise-memory-exns=yes/g' <run-tests.php >r
+		rm -f run-tests.php
+		cp r run-tests.php
+	fi
+fi
+export TEST_PHP_ARGS
+
+make test
 
 for i in tests/*.diff tests/*.mem
 	do
@@ -42,38 +53,36 @@ return $_ret
 
 #----------------
 
-ret=0
+export USE_VALGRIND
 
-export NO_INTERACTION=1
+BASE=`pwd`
+export BASE
+
+ret=0
 
 # Xdebug versions < 2.3.3 causes PHP to crash when enabled with PCS ext (see
 # http://bugs.xdebug.org/view.php?id=1151).
 # **TODO: To be confirmed on PCS
 # Note that xdebug is not pre-installed on every PHP version (so, phpenv may fail)
 
-phpenv config-rm xdebug.ini || :
+if [ -n "$TRAVIS" ] ; then
+	phpenv config-rm xdebug.ini || :
+fi
 
-valgrind --version
+#-- First, run base tests, then examples
 
-#-- First, run base dir tests
-
-run_tests
-[ $? = 0 ] || ret=$?
-
-#-- Then, test the client extensions
-
-make install
-
-cd examples/pcs_test
-
-run_tests
-[ $? = 0 ] || ret=$?
-
-cd ../example1
-
-run_tests
-[ $? = 0 ] || ret=$?
+for d in . examples/example1 examples/pcs_test
+do
+	run_tests $d
+	[ $? = 0 ] || ret=$?
+done
 
 #----
+
+if [ "$ret" = 0 ] ;then
+	echo "======== SUCCESS ============="
+else
+	echo "************ One or more errors were encoutered **************"
+fi
 
 exit $ret
